+++ osidb/__init__.py
@@ -1,5 +1,5 @@
 """
     osidb version
 """
-__version__ = "4.5.2"
+__version__ = "4.5.3"
 VERSION = __version__