+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Implement resolution steps for duplicate tracker validation (OSIDB-3588)
+
 ### Changed
 - Add upstream references to Jira trackers on creation (OSIDB-3148)
 - Change ACL mixin serializer to support internal ACLs (OSIDB-3578)